+++ src/AssemblyVersioning.cs
@@ -24,5 +24,5 @@
 //   2) Update the installer version to match the AssemblyVersion below.
 //   These can be found in SqlBuildManager.Setup -> Organize Your Setup -> General Information
 
-[assembly: AssemblyVersion("14.3.0")]
-[assembly: AssemblyFileVersion("14.3.0")]
+[assembly: AssemblyVersion("14.3.1")]
+[assembly: AssemblyFileVersion("14.3.1")]